crowdloan list structure refactored
